Add js code to populate html with weather data
index.js updates the content on the page, the only interactivity it adds is for the search function.
data.js was updated to include the weather icons and now formats the wind direction and precipitation type data to be more readable.
styles.css was updated to include the new icons, it will recieve a more thorough update in the future.
index.html was updated to include the new icons and more daily and hourly data.
Links to the other pages were removed from index.html as they are not yet implemented.